gcp/requirements.txt
pillow==10.1.0
tensorflow==2.13.0
google-cloud-storage==1.16.1
